+++ assets/README.md
@@ -1,0 +1,1 @@
+フロントエンドで使用される静的ファイル置き場。
