+++ components/README.md
@@ -1,0 +1,1 @@
+Reactコンポーネント置き場。
