+++ testdata/README.md
@@ -1,0 +1,1 @@
+テストで使用されるファイル置き場。
