+++ worker/README.md
@@ -1,0 +1,1 @@
+service workerで使用されるコード置き場。
